5 ok

+++ src/main/kotlin/ar/edu/unahur/obj2/impostoresPaises/Pais.kt
@@ -3,7 +3,7 @@ package ar.edu.unahur.obj2.impostoresPaises
 class Pais(
         val nombre :String,
         val codigoISo3 : String,
-        val poblacion : Int,
+        val poblacion : Long,
         val continente : String,
         val bloqueRegional : List<String>,
         val idomasOficiales : List<String>
